spaceship designer: begin get segments of interest

+++ src/maths/grid.hpp
@@ -45,10 +45,10 @@ grid_space_to_world_space(const glm::ivec2 pos, const int grid_size)
 }
 
 [[nodiscard]] inline glm::ivec2
-world_space_to_grid_space(const glm::vec2& pos, const float grid_size)
+world_space_to_grid_space(const glm::vec2& pos, const int grid_size)
 {
-  const int grid_x = static_cast<int>(glm::abs(pos.x) / grid_size);
-  const int grid_y = static_cast<int>(glm::abs(pos.y) / grid_size);
+  const int grid_x = static_cast<int>(glm::abs(pos.x) / static_cast<float>(grid_size));
+  const int grid_y = static_cast<int>(glm::abs(pos.y) / static_cast<float>(grid_size));
   if (pos.x >= 0 && pos.y >= 0)
     return { grid_x, grid_y };
